==> 11_FeatureEngineering.ipynb <==
# Feature Engineering

## 1. Setup

import pandas as pd
import numpy as np
from summarytools import dfSummary
from sklearn.preprocessing import MinMaxScaler

from env_setup import *
from functions.t_test import *

pd.set_option('display.max_columns', None)

# Load data
df_raw = pd.read_csv(fr"{dataout}//{dataset}_EDA.csv")
df_raw.head()
df_ori = df_raw.copy()

## 2. Feature Engineering

# Split categorical and numeric
l_cols_cat = [i for i in df_ori if df_ori[i].dtype==object]
l_cols_num = [i for i in df_ori if df_ori[i].dtype!=object and i != 'Churned']

print(f"Categorical columns: {', '.join(l_cols_cat)}")
print(f"Numeric columns: {', '.join(l_cols_num)}")

### 2.1. Categorical One Hot Encoding
Apply one-hot encoding to categorical variables for modelling

# Summary Statistics for Categorical
## Encode categorical
df_ori_encoded = pd.get_dummies(df_ori, columns=l_cols_cat)
l_cols_encoded = [i for i in df_ori_encoded.columns if any(x in i for x in l_cols_cat)]

for col in l_cols_encoded:
    df_ori_encoded[col] = np.where(df_ori_encoded[col]==True, 1, 0)

stats_cat = df_ori_encoded.groupby('Churned')[l_cols_encoded].sum().reset_index()

## Scale
stats_cat_scaled = stats_cat.copy()
for col in l_cols_encoded:
    stats_cat_scaled[col] = np.where(stats_cat_scaled['Churned']==0, 
                                    stats_cat_scaled[col] / float(df_ori_encoded[df_ori_encoded['Churned']==0]['Churned'].count()),
                                    stats_cat_scaled[col] / float(df_ori_encoded[df_ori_encoded['Churned']==1]['Churned'].count())
    )
stats_cat_scaled

### 2.2. Numerical Binning
Apply binning to numerical variables.

Mostly not necessary as bands are not distinct enough but just in case needed for modelling.

# Create Bins
df_ori_binned = df_ori.copy()

d_bins = {
    'Age': list(range(0, 80, 10)), 
    'Income': list(range(0, 200_000, 10_000)), 
    'Tenure': [0, 5], # Hypothesis test cut off = 5
    #'TransactionDate': list(range(0, 390, 30)),
    #'TransactionAmount': list(range(0, 600, 100)), 
    # 'NumSupportCalls': [], # Not required - too small and no major difference in hypothesis test
    # 'NumComplaints': [] # Not required - too small and no major difference in hypothesis test
}

for i,j in d_bins.items():
    df_ori_binned[f'{i}_bin'] = pd.cut(df_ori_binned[i], bins=j)

# Fix nans
df_ori_binned['Tenure_bin'] = df_ori_binned['Tenure_bin'].cat.add_categories('>5')
df_ori_binned.loc[df_ori_binned['Tenure_bin'].isna(), 'Tenure_bin'] = '>5'

# View Statistics
dfSummary(df_ori_binned[[f'{i}_bin' for i in d_bins.keys()]])

# One Hot Encoding
df_ori_encoded_num = pd.get_dummies(df_ori_binned[[f'{i}_bin' for i in d_bins.keys()]], columns=[f'{i}_bin' for i in d_bins.keys()])

for col in df_ori_encoded_num.columns:
    df_ori_encoded_num[col] = np.where(df_ori_encoded_num[col]==True, 1, 0)

df_ori_encoded_num.head()

for col in df_ori_encoded_num.columns:
    df_ori_encoded_num[col] = np.where(df_ori_encoded_num[col]==True, 1, 0)

df_ori_encoded_label = pd.concat([df_ori_encoded_num, df_ori['Churned']], axis=1)
stats_bin = df_ori_encoded_label.groupby('Churned')[df_ori_encoded_num.columns].sum().reset_index()

## Scale
stats_bin_scaled = stats_bin.copy()
for col in df_ori_encoded_num.columns:
    stats_bin_scaled[col] = np.where(stats_bin_scaled['Churned']==0, 
                                    stats_bin_scaled[col] / float(df_ori_encoded[df_ori_encoded['Churned']==0]['Churned'].count()),
                                    stats_bin_scaled[col] / float(df_ori_encoded[df_ori_encoded['Churned']==1]['Churned'].count())
    )
stats_bin_scaled

## 3. Normalisation
Normalisation used since most variables are not normally distributed

scaler = MinMaxScaler() 
scaled_values = scaler.fit_transform(df_raw[l_cols_num]) 
df_ori_norm = pd.DataFrame(scaled_values, columns=[f'{i}_norm' for i in l_cols_num])

## 4. Recombine

df_ori_encoded.columns

df_fe = pd.concat(
    [
        df_raw, # Original
        df_ori_norm, # Normalised
        df_ori_encoded.drop(columns=['Age', 'Income', 'Tenure', 'NumSupportCalls', 'NumComplaints',
       'Churned', 'Purchase', 'Refund', 'Subscription Renewal', 'Support Fee',
       'Upgrade', 'txn_mean', 'txn_count']), # Categorical dummy
    ],
    axis = 1)
df_fe.head()

## 5. Export

# Export datasets
df_fe.to_csv(fr"{dataout}//{dataset}_FE.csv", index=False)
