==> 20_Modelling.ipynb <==
# Modelling

## 1. Setup

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from summarytools import dfSummary

from env_setup import *

pd.set_option('display.max_columns', None)

# Load data
X_test = pd.read_csv(fr"{dataout}//{dataset}_test.csv")
y_test = X_test[['Churned']]
X_test.drop(columns=['Churned'], inplace=True)

X_train = pd.read_csv(fr"{dataout}//{dataset}_train.csv")
y_train = X_train[['Churned']]
X_train.drop(columns=['Churned'], inplace=True)

X_train_sm = pd.read_csv(fr"{dataout}//{dataset}_SMOTE_train.csv")
y_train_sm = X_train_sm[['Churned']]
X_train_sm.drop(columns=['Churned'], inplace=True)

X_train_smenn = pd.read_csv(fr"{dataout}//{dataset}_SMOTEENN_train.csv")
y_train_smenn = X_train_smenn[['Churned']]
X_train_smenn.drop(columns=['Churned'], inplace=True)

X_train.columns

# Original vs Binned columns
l_cols_num = [
    'Age', 'Income', 'Tenure', 'NumSupportCalls', 'NumComplaints',
    'Purchase', 'Refund', 'Subscription Renewal', 'Support Fee', 'Upgrade',
    'txn_mean', 'txn_count'
    ]

l_cols_cat = [
    #'Gender_Female', 
    'Gender_Male', 
    'Location_Rural', 'Location_Suburban', 
    #'Location_Urban',
    ]

l_cols_norm = [
    'Age_norm', 'Income_norm', 'Tenure_norm',
    'NumSupportCalls_norm', 'NumComplaints_norm', 'Purchase_norm',
    'Refund_norm', 'Subscription Renewal_norm', 'Support Fee_norm',
    'Upgrade_norm', 'txn_mean_norm', 'txn_count_norm'
    ]

## 2. Baseline Models

### 2.1. Logistic Regression

import pandas as pd
import statsmodels.api as sm 
from sklearn.metrics import classification_report, f1_score

class logistic_regression:
    def __init__(self, y_train, x_train, cols, name, y=y_test, x=X_test):
        self.name = name
        self.y_train = y_train
        self.x_train = x_train
        self.y = y_test
        self.x = X_test
        self.cols = cols
        self.model = sm.Logit(self.y_train, self.x_train[cols]).fit()

    def summary(self):
        print(self.model.summary())

    def predict(self, confusion=False):
        yhat = self.model.predict(self.x[self.cols]) 
        prediction = list(map(round, yhat))
        self.cm = classification_report(self.y, prediction)
        f1 = f1_score(self.y, prediction)
        self.f1 = f1
        print(f"{self.name}: F1 = {f1}")    
        if confusion:
            print ("Confusion Matrix : \n", self.cm)

    def feature_selection(self):
        # Remove columns with no statistical differences
        try:
            self.cols.remove('Age')
            self.cols.remove('TransactionDate')
            self.cols.remove('NumComplaints')
        except:
            pass
        self.model = sm.Logit(self.y_train, self.x_train[self.cols]).fit()
        self.predict()
        f1 = self.f1

        for i in self.cols:
            self.cols.remove(i)
            self.model = sm.Logit(self.y_train, self.x_train[self.cols]).fit()
            print(self.cols)
            self.predict()
            new_f1 = self.f1

            if new_f1 < f1:
                self.cols.append(i)

ori_baseLogit_num = logistic_regression(y_train, X_train, l_cols_num+l_cols_cat, 'ori_baseLogit_num')
ori_baseLogit_num.predict()
ori_baseLogit_norm = logistic_regression(y_train, X_train, l_cols_norm+l_cols_cat, 'ori_baseLogit_norm')
ori_baseLogit_norm.predict()
print('\n')
sm_baseLogit_num = logistic_regression(y_train_sm, X_train_sm, l_cols_num+l_cols_cat, 'sm_baseLogit_num')
sm_baseLogit_num.predict()
sm_baseLogit_norm = logistic_regression(y_train_sm, X_train_sm, l_cols_norm+l_cols_cat, 'sm_baseLogit_norm')
sm_baseLogit_norm.predict()
print('\n')
smenn_baseLogit_num = logistic_regression(y_train_smenn, X_train_smenn, l_cols_num+l_cols_cat, 'smenn_baseLogit_num')
smenn_baseLogit_num.predict()
smenn_baseLogit_norm = logistic_regression(y_train_smenn, X_train_smenn, l_cols_norm+l_cols_cat, 'smenn_baseLogit_norm')
smenn_baseLogit_norm.predict()

### 2.2. Decision Trees

import pandas as pd
from sklearn import tree
from sklearn.tree import DecisionTreeClassifier
from sklearn.metrics import classification_report, f1_score
from sklearn.model_selection import GridSearchCV

class decision_tree:
    def __init__(self, y_train, x_train, cols, name, y=y_test, x=X_test):
        self.name = name
        self.y_train = y_train
        self.x_train = x_train
        self.y = y_test
        self.x = X_test
        self.cols = cols
        self.model = DecisionTreeClassifier().fit(self.x_train[self.cols], self.y_train)

    def summary(self):
        tree.plot_tree(self.model, feature_names=self.cols)

    def predict(self, confusion=False):
        yhat = self.model.predict(self.x[self.cols]) 
        prediction = list(map(round, yhat))
        self.cm = classification_report(self.y, prediction)
        f1 = f1_score(self.y, prediction)
        self.f1 = f1
        print(f"{self.name}: F1 = {f1}")    
        if confusion:
            print ("Confusion Matrix : \n", self.cm)

    def feature_selection(self):
        # Remove columns with no statistical differences
        try:
            self.cols.remove('Age')
            self.cols.remove('TransactionDate')
            self.cols.remove('NumComplaints')
        except:
            pass
        self.model = DecisionTreeClassifier().fit(self.x_train[self.cols], self.y_train)
        self.predict()
        f1 = self.f1

        for i in self.cols:
            self.cols.remove(i)
            self.model = DecisionTreeClassifier().fit(self.x_train[self.cols], self.y_train)
            print(self.cols)
            self.predict()
            new_f1 = self.f1

            if new_f1 < f1:
                self.cols.append(i)

    def fine_tune(self, params):
        random_search = GridSearchCV(
            DecisionTreeClassifier(),
            params, 
            n_jobs=-1, 
            cv=5, 
            scoring='f1_weighted')
        
        random_search.fit(self.x_train[self.cols], self.y_train)
        self.best_params_random = random_search.best_params_
        self.best_score_random = random_search.best_score_
        self.model = random_search.best_estimator_
        self.predict()

# Normalisation not required for decision trees
ori_baseDT_num = decision_tree(y_train, X_train, l_cols_num+l_cols_cat, 'ori_baseDT_num')
ori_baseDT_num.predict()
print('\n')
sm_baseDT_num = decision_tree(y_train_sm, X_train_sm, l_cols_num+l_cols_cat, 'sm_baseDT_num')
sm_baseDT_num.predict()
print('\n')
smenn_baseDT_num = decision_tree(y_train_smenn, X_train_smenn, l_cols_num+l_cols_cat, 'smenn_baseDT_num')
smenn_baseDT_num.predict()

### 2.3. Naive Bayes

import pandas as pd
from sklearn.naive_bayes import GaussianNB
from sklearn.metrics import classification_report, f1_score

class naive_bayes:
    def __init__(self, y_train, x_train, cols, name, y=y_test, x=X_test):
        self.name = name
        self.y_train = y_train.values.ravel()
        self.x_train = x_train
        self.y = y_test.values.ravel()
        self.x = X_test
        self.cols = cols
        self.model = GaussianNB().fit(self.x_train[self.cols].values, self.y_train)

    def predict(self, confusion=False):
        yhat = self.model.predict(self.x[self.cols].values) 
        prediction = list(map(round, yhat))
        self.cm = classification_report(self.y, prediction)
        f1 = f1_score(self.y, prediction)
        self.f1 = f1
        print(f"{self.name}: F1 = {f1}")    
        if confusion:
            print ("Confusion Matrix : \n", self.cm)

    def feature_selection(self):
        # Remove columns with no statistical differences
        try:
            self.cols.remove('Age')
            self.cols.remove('TransactionDate')
            self.cols.remove('NumComplaints')
        except:
            pass
        self.model = GaussianNB().fit(self.x_train[self.cols].values, self.y_train)
        self.predict()
        f1 = self.f1

        for i in self.cols:
            self.cols.remove(i)
            self.model = GaussianNB().fit(self.x_train[self.cols].values, self.y_train)
            print(self.cols)
            self.predict()
            new_f1 = self.f1

            if new_f1 < f1:
                self.cols.append(i)

ori_baseNB_num = naive_bayes(y_train, X_train, l_cols_num+l_cols_cat, 'ori_baseNB_num')
ori_baseNB_num.predict()
ori_baseNB_norm = naive_bayes(y_train, X_train, l_cols_norm+l_cols_cat, 'ori_baseNB_norm')
ori_baseNB_norm.predict()
print('\n')
sm_baseNB_num = naive_bayes(y_train_sm, X_train_sm, l_cols_num+l_cols_cat, 'sm_baseNB_num')
sm_baseNB_num.predict()
sm_baseNB_norm = naive_bayes(y_train_sm, X_train_sm, l_cols_norm+l_cols_cat, 'sm_baseNB_norm')
sm_baseNB_norm.predict()
print('\n')
smenn_baseNB_num = naive_bayes(y_train_smenn, X_train_smenn, l_cols_num+l_cols_cat, 'smenn_baseNB_num')
smenn_baseNB_num.predict()
smenn_baseNB_norm = naive_bayes(y_train_smenn, X_train_smenn, l_cols_norm+l_cols_cat, 'smenn_baseNB_norm')
smenn_baseNB_norm.predict()

### 2.4. Random Forest

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report, f1_score

class random_forest:
    def __init__(self, y_train, x_train, cols, name, y=y_test, x=X_test):
        self.name = name
        self.y_train = y_train
        self.x_train = x_train[cols]
        self.y = y_test
        self.x = X_test[cols]
        self.cols = cols
        self.model = RandomForestClassifier(n_estimators = 10).fit(self.x_train[cols], self.y_train)

    def summary(self):
        tree.plot_tree(self.model, feature_names=self.cols)

    def predict(self, confusion=False):
        yhat = self.model.predict(self.x[self.cols]) 
        prediction = list(map(round, yhat))
        self.cm = classification_report(self.y, prediction)
        f1 = f1_score(self.y, prediction)
        self.f1 = f1
        print(f"{self.name}: F1 = {f1}")    
        if confusion:
            print ("Confusion Matrix : \n", self.cm)

# Normalisation and Balancing not required for decision trees
ori_baseRF_num = random_forest(y_train_sm, X_train_sm, l_cols_num+l_cols_cat, 'ori_baseRF_num')
ori_baseRF_num.predict()

## 3. Fine-Tuning

# Feature selection - LR
smenn_FT_Logit_num = logistic_regression(y_train_smenn, X_train_smenn, l_cols_num+l_cols_cat, 'smenn_FT_Logit_num')
smenn_FT_Logit_num.predict()
smenn_FT_Logit_num.feature_selection()
smenn_FT_Logit_num.feature_selection()

# Feature selection - DT
ori_FS_DT_num = decision_tree(y_train_sm, X_train_sm, l_cols_num+l_cols_cat, 'sm_baseDT_num')
ori_FS_DT_num.predict()
ori_FS_DT_num.feature_selection()
ori_FS_DT_num.feature_selection()

# Feature selection - NB
sm_FT_NB_num = naive_bayes(y_train_sm, X_train_sm, l_cols_num+l_cols_cat, 'sm_FT_NB_num')
sm_FT_NB_num.predict()
sm_FT_NB_num.feature_selection()
sm_FT_NB_num.feature_selection()

### Best Models

# Logistic Regression
LR_model = logistic_regression(y_train_smenn, X_train_smenn, ['NumComplaints', 'Gender_Male', 'NumSupportCalls', 'Tenure'], 'LR_model')
LR_model.predict(confusion=True)

LR_model.summary()

# Decision Tree
DT_model = decision_tree(y_train_smenn, X_train_smenn, ['NumComplaints', 'Gender_Male', 'Tenure', 'Refund'], 'DT_model')
DT_model.predict(confusion=True)

# Naive Bayes
NB_model = naive_bayes(y_train_sm, X_train_sm, ['NumComplaints', 'Gender_Male', 'Tenure'], 'NB_model')
NB_model.predict(confusion=True)

# Define the parameter distribution to sample from
param_dist = {'criterion':['gini','entropy'],
              'max_depth':np.arange(1,10).tolist()[0::2],
              'min_samples_split':np.arange(100,200).tolist()[0::2],
              'max_leaf_nodes':np.arange(3,10).tolist()[0::2]}

ori_FT_DT_num = decision_tree(y_train, X_train, ['NumSupportCalls', 'TransactionType_Refund', 'Income', 'Tenure', 'Location_Rural'], 'ori_FT_DT_num')
ori_FT_DT_num.fine_tune(param_dist)

print(ori_FT_DT_num.best_params_random)
print('\n')
print(ori_FT_DT_num.best_score_random)
